Fix up boolean option handling to be more explicit, and error on bad values.
When dealing with a tool that can overwrite files, my philosophy is to only
allow well known boolean argument values to be used.  The selector method I'm replacing
used a default setting to pass through any non-true values (such as a
mispelling of "true", which would silently become 'false').  I decided to
tighten it up to require actual 'true' or 'false' values to be specified, and
fail to execute otherwise.

In addition, I allow the common puppet aliases for 'true' (ie. 'yes' and 'on'),
or 'false' ('no' and 'off'), since they make sense in this context.

note -- boolean handling in puppet is kinda screwy...